it looks like we have a working interpolation system, mpp processing, and a working framework to clalculate max GM from an etas input. started a notebook to work out and outline the model(s) to transform ETAS rates to source event magnitudes.
